inital move of clearly assignable  functions
